refactor reportmakedecisionmodal into own component
